break down committee section further for readability
